[14.0][MIG] account_asset_non_deductible: Migration to 14.0
