+++ account_asset_non_deductible/models/__init__.py
@@ -2,4 +2,4 @@
 # Eric Antones <eantones@nuobit.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl)
 
-from . import account_invoice
+from . import account_move
