--- setup/account_asset_non_deductible/setup.cfg
@@ -1,2 +1,0 @@
-[bdist_wheel]
-universal=1
